refactor commandmap
